Refactor ApiResponse & change folder name and namespace

+++ src/DevnotMentor.WebAPI/Controllers/BaseController.cs
@@ -6,9 +6,9 @@ namespace DevnotMentor.WebAPI.Controllers
     public class BaseController : ControllerBase
     {
         [NonAction]
-        protected IActionResult ApiResponse(ApiResponse response) => StatusCode((int)response.StatusCode, response);
+        protected IActionResult ApiResponse(ApiResponse response) => StatusCode(response.StatusCode, response);
 
         [NonAction]
-        protected IActionResult ApiResponse<T>(ApiResponse<T> response) => StatusCode((int)response.StatusCode, response);
+        protected IActionResult ApiResponse<T>(ApiResponse<T> response) => StatusCode(response.StatusCode, response);
     }
 }
